MAINT: Move homs_system to config
